pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "faker>=40.1.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "strawberry-graphql[fastapi]>=0.288.3",
    "uvicorn>=0.40.0",
]
